feat: add auth middleware to controllers
